chore: rename mfa to user management
