+++ app/page.tsx
@@ -3,10 +3,10 @@ import {usePrivy} from '@privy-io/react-auth';
 
 import Login from './components/Login';
 import Wallets from './components/Wallets';
-import MFA from './components/MFA';
+import UserManagement from './components/UserManagement';
 
 export default function Home() {
-  const {ready} = usePrivy();
+  const {ready, authenticated} = usePrivy();
   if (!ready) {
     return <div></div>;
   }
@@ -46,9 +46,11 @@ export default function Home() {
         <div className="w-full md:w-1/2">
           <Wallets />
         </div>
-        <div className="w-full md:w-1/2">
-          <MFA />
-        </div>
+        {authenticated && (
+          <div className="w-full md:w-1/2">
+            <UserManagement />
+          </div>
+        )}
       </div>
     </div>
   );
